src/agentic_platform/agent/my_new_multi_agent_chat/requirements.txt
fastapi==0.115.12
uvicorn==0.34.0
boto3==1.37.27
pyjwt==2.10.1
cryptography==44.0.2
requests
litellm>=1.66.2
pydantic>=2.10.6
strands-agents[litellm,openai]>=0.1.6
strands-agents-tools>=0.1.9
aws-opentelemetry-distro>=0.10.1
